refactor(ka): move oauth2_credentials from pkg/ to internal/ (#995)
oauth2_credentials.go depends on internal/kubernautagent/config for
OAuth2Config. Since no pkg/ code uses NewOAuth2ClientCredentialsTransport
(only cmd/ does), move it to internal/kubernautagent/llm/transport/.

Eliminates the last pkg/kubernautagent -> internal/kubernautagent
layering violation.
