+++ internal/kubernautagent/llm/transport/oauth2_credentials.go
